# docs/swagger.yaml
openapi: 3.0.3
security:
  - BearerAuth: []
info:
  title: AI-Persona Backend API
  version: 1.0.0
  description: |
    # AI-Persona SaaS Backend API - Core Authentication & User Management

    ## **Overview**
    This API provides enterprise-grade authentication and user management for the AI-Persona SaaS platform. The core module is designed for production environments with advanced security features and compliance.

    ## **Core Features**
    - **Authentication**: Local (email/password) and Google OAuth authentication
    - **Account Lifecycle**: PENDING_VERIFY → ACTIVE → DEACTIVATED → PENDING_DELETION
    - **Email Verification**: 24-hour TTL verification tokens with resend capability
    - **Password Reset**: Secure token-based password reset with 1-hour TTL
    - **Session Management**: Device tracking, session revocation, and token rotation
    - **Account Lockout**: Automatic lockout after failed login attempts (5 attempts → 15 min)
    - **GDPR Compliance**: Account deletion with 30-day grace period
    - **Webhook Integration**: External service integration for persona trait updates

    ## **Security Architecture**
    - **JWT Security**: RSA-256 signing with key rotation and JWKS endpoint
    - **Password Security**: Strength validation + HIBP breach detection
    - **Account Protection**: Rate limiting, lockout, session management
    - **Input Validation**: Comprehensive sanitization and validation
    - **Audit Trail**: Complete event logging with IP, user agent, trace IDs

...
    ```
...

...

...

...

...
servers:
...
paths:
...
  /api/conversations:
    get:
      tags:
        - Conversations
      summary: List user conversations
      description: |
        Get a list of conversations for the authenticated user.
        Returns user's private conversations plus all shared conversations in their workspace.
      security:
        - BearerAuth: []
      parameters:
        - name: archived
          in: query
          description: Include archived conversations
          required: false
          schema:
            type: boolean
            example: false
      responses:
        "200":
          description: Conversations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Conversations retrieved successfully"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/conversations/{id}/visibility:
    patch:
      tags:
        - Conversations
      summary: Update conversation visibility
      description: |
        Update conversation visibility to PRIVATE or SHARED.
        Only conversation owner or workspace admin can change visibility.
        Cannot change visibility of archived conversations.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
            format: uuid
            example: "conversation_id_here"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - visibility
              properties:
                visibility:
                  type: string
                  enum: [PRIVATE, SHARED]
                  example: "SHARED"
      responses:
        "200":
          description: Conversation visibility updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Conversation visibility updated successfully"
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "conversation_id_here"
                      visibility:
                        type: string
                        enum: [PRIVATE, SHARED]
                        example: "SHARED"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Cannot change visibility of archived conversation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "Cannot change visibility of archived conversation"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/conversations/{id}/files:
    post:
      tags:
        - Files
      summary: Request file upload URL
      description: |
        Request a presigned upload URL for file attachment.
        Only conversation owner or workspace member can upload files.
        Supports images and PDFs up to 10 MB.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
            format: uuid
            example: "conversation_id_here"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - filename
                - mimeType
                - sizeBytes
              properties:
                filename:
                  type: string
                  maxLength: 200
                  example: "document.pdf"
                mimeType:
                  type: string
                  pattern: "^(image/|application/pdf)$"
                  example: "application/pdf"
                sizeBytes:
                  type: integer
                  minimum: 1
                  maximum: 10485760
                  example: 1024000
      responses:
        "200":
          description: File upload URL generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "File upload URL generated successfully"
                  data:
                    type: object
                    properties:
                      presignedUrl:
                        type: string
                        format: uri
                        example: "https://example.com/uploads/file_id"
                      fileId:
                        type: string
                        example: "file_id_here"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "File size must be less than 10 MB"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/conversations/{id}/archive:
    patch:
      tags:
        - Conversations
      summary: Archive or unarchive conversation
      description: |
        Archive or unarchive a conversation.
        Only conversation owner or workspace admin can archive.
        Archived conversations cannot be edited, receive new messages, or change visibility.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
            format: uuid
            example: "conversation_id_here"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - archived
              properties:
                archived:
                  type: boolean
                  example: true
      responses:
        "200":
          description: Conversation archive status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Conversation archived successfully"
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "conversation_id_here"
                      archived:
                        type: boolean
                        example: true
                      archivedAt:
                        type: string
                        format: date-time
                        example: "2025-08-05T08:40:00.000Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Conversation cannot be archived
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "Conversation is already archived"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/conversations/{id}/share:
    post:
      tags:
        - Conversations
      summary: Create or refresh shareable link
      description: |
        Create or refresh a shareable link for a conversation.
        Only SHARED conversations can have shareable links.
        Archived conversations cannot be shared.
        Only conversation owner or workspace admin can share.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
            format: uuid
            example: "conversation_id_here"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresInDays:
                  type: integer
                  minimum: 1
                  maximum: 365
                  default: 30
                  description: Days until link expires
                  example: 30
      responses:
        "200":
          description: Shareable link created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Shareable link created successfully"
                  data:
                    type: object
                    properties:
                      url:
                        type: string
                        format: uri
                        example: "https://app.com/p/abc123def456"
                      expiresAt:
                        type: string
                        format: date-time
                        example: "2025-09-04T08:40:00.000Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Conversation cannot be shared
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "Only SHARED conversations can have shareable links"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Conversation:
      type: object
      properties:
        id:
          type: string
          example: "conversation_id_here"
        title:
          type: string
          example: "Chat with AI Assistant"
        persona:
          type: object
          properties:
            id:
              type: string
              example: "persona_id_here"
            name:
              type: string
              example: "AI Assistant"
            avatarUrl:
              type: string
              format: uri
              example: "https://example.com/avatar.jpg"
        owner:
          type: object
          properties:
            id:
              type: string
              example: "user_id_here"
            name:
              type: string
              example: "John Doe"
            email:
              type: string
              format: email
              example: "john@example.com"
        visibility:
          type: string
          enum: [PRIVATE, SHARED]
          example: "PRIVATE"
        archivedAt:
          type: string
          format: date-time
          nullable: true
          example: null
        lastMessage:
          type: string
          example: "Hello! How can I help you today?"
        messageCount:
          type: integer
          example: 10
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Bad request"
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: "email"
                    message:
                      type: string
                      example: "Invalid email format"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Authentication required"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Insufficient permissions"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Resource not found"
...
    TooManyRequests:
      description: Too Many Requests
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Too many requests, please try again later"
...